Remove Codex compatibility references
